LayoutTests/streams/reference-implementation/bad-underlying-sources-expected.txt


FAIL Underlying source start: throwing getter assert_throws: constructing the stream should re-throw the error function "function () {
        new ReadableStream({
            ge..." threw object "TypeError: ReadableStream constructor object start proper..." ("TypeError") expected object "Error: a unique string" ("Error")
PASS Underlying source start: throwing method 
TIMEOUT Underlying source: throwing pull getter (initial pull) Test timed out
TIMEOUT Underlying source: throwing pull method (initial pull) Test timed out
FAIL Underlying source: throwing pull getter (second pull) read is not implemented
FAIL Underlying source: throwing pull method (second pull) read is not implemented
FAIL Underlying source: throwing cancel getter cancel is not implemented
FAIL Underlying source: throwing cancel method cancel is not implemented
FAIL Underlying source: throwing strategy getter assert_throws: enqueue should throw the error function "function () { c.enqueue('a'); }" did not throw
FAIL Underlying source: throwing strategy.size getter assert_throws: enqueue should throw the error function "function () { c.enqueue('a'); }" did not throw
FAIL Underlying source: throwing strategy.size method assert_throws: enqueue should throw the error function "function () { c.enqueue('a'); }" did not throw
FAIL Underlying source: throwing strategy.shouldApplyBackpressure getter assert_throws: enqueue should throw the error function "function () { c.enqueue('a'); }" did not throw
FAIL Underlying source: throwing strategy.shouldApplyBackpressure method assert_throws: enqueue should throw the error function "function () { c.enqueue('a'); }" did not throw
FAIL Underlying source: strategy.size returning NaN assert_unreached: enqueue didn't throw Reached unreachable code
FAIL Underlying source: strategy.size returning -Infinity assert_unreached: enqueue didn't throw Reached unreachable code
FAIL Underlying source: strategy.size returning +Infinity assert_unreached: enqueue didn't throw Reached unreachable code
FAIL Underlying source: calling close twice on an empty stream should throw the second time assert_throws: second call to close should throw a TypeError function "function () {
    [native code]
}" did not throw
FAIL Underlying source: calling close twice on a non-empty stream should throw the second time assert_throws: second call to close should throw a TypeError function "function () {
    [native code]
}" did not throw
FAIL Underlying source: calling close on an empty canceled stream should not throw cancel is not implemented
FAIL Underlying source: calling close on a non-empty canceled stream should not throw cancel is not implemented
FAIL Underlying source: calling close after error should throw assert_throws: call to close should throw a TypeError function "function () {
    [native code]
}" did not throw
FAIL Underlying source: calling error twice should throw the second time assert_throws: second call to error should throw a TypeError function "function () {
    [native code]
}" did not throw
FAIL Underlying source: calling error after close should throw assert_throws: call to error should throw a TypeError function "function () {
    [native code]
}" did not throw
